Extract mock of HttpClientEngine to :data:users:network-test
Thanks to that change I would be able to use it in E2E tests.
