+++ data/users/network-test/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("kotlin")
+}
+
+dependencies {
+    api(libs.ktor.core)
+    implementation(project(":data:users:network"))
+    implementation(libs.ktor.test)
+}
