+++ data/users/network-test/src/main/kotlin/dev/burnoo/demo/listapp/data/users/network/test/GetUsersResponse.kt
@@ -1,4 +1,4 @@
-package dev.burnoo.demo.listapp.data.users.network.utils
+package dev.burnoo.demo.listapp.data.users.network.test
 
 fun getUsersResponse(firstUserId: String) = """
 {
