+++ settings.gradle.kts
@@ -30,3 +30,4 @@ include(":data:users:core")
 include(":data:users:core-test")
 include(":data:users:model")
 include(":data:users:network")
+include(":data:users:network-test")
